refactor(encoding): extract HcMatchGenerator to hc/generator.rs
Move the HashChain / binary-tree match generator out of match_generator.rs into
hc/generator.rs: the HcMatchGenerator storage + methods, the btlazy2 cost
helpers, the optimal-plan body macros, and the build_optimal_plan driver. The
cross-module macros (bt_insert_*, hash3_candidate, for_each_repcode) now resolve
through hc::generator; their consumers (match_table::storage, hc) are repointed.

Verbatim move (only relative paths rewritten to absolute and the driver-facing
surface widened to pub(crate)), so the output is byte-identical. The small
HcBackend enum stays in match_generator (its impl moved). match_generator.rs:
11306 -> 7438.

Part of #441.
